python-server/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Career Path Simulator - AI-driven multi-agent system for personalized career planning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "pypdf>=6.4.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
